Avoid servicing conflicts between MBeanFactory and other channels
Third in a series of patches aimed at allowing parallel requests to the
Manager application to deploy different applications in parallel rather
than using a sync block to deploy them serially.
Prior to this patch MBeanFactory did not check whether the app to be
deployed was already listed as being serviced. This is now checked
(using the new thread safe tryAddServiced) before trying to add or
remove an app via MBeanFactory